+++ ci/ubuntu/build.sh
@@ -47,9 +47,12 @@ python -m http.server &> /dev/null &
 echo "PHP version"
 php -v
 PHPVersionMinor=$(php --version | head -n 1 | cut -d " " -f 2 | cut -c 1,3)
+if [ ${PHPVersionMinor} -gt 83 ]; then
+    cd php && curl -LO https://phar.phpunit.de/phpunit-13.phar
 if [ ${PHPVersionMinor} -ge 83 ]; then
     cd php && curl -LO https://phar.phpunit.de/phpunit-12.phar
     echo "PHPUnit version"
+    php phpunit-13.phar --version
     php phpunit-12.phar --version
 else
     cd php && curl -LO https://phar.phpunit.de/phpunit-10.phar