+++ src/mapuvraster.cpp
@@ -60,39 +60,45 @@
 typedef struct {
 
   /* query cache results */
-  int query_results;
+  int query_results = 0;
 
-  int refcount;
+  int refcount = 0;
 
-  float *u; /* u values */
-  float *v; /* v values */
-  int width;
-  int height;
-  rectObj extent;
-  int next_shape;
+  float *u = nullptr; /* u values */
+  float *v = nullptr; /* v values */
+  int width = 0;
+  int height = 0;
+  rectObj extent{};
+  int next_shape = 0;
 
   /* To improve performance of GetShape() when queried on increasing shapeindex
    */
-  long last_queried_shapeindex; // value in [0, query_results[ range
-  size_t last_raster_off;       // value in [0, width*height[ range
+  long last_queried_shapeindex = 0; // value in [0, query_results[ range
+  size_t last_raster_off = 0;       // value in [0, width*height[ range
 
-  bool needsLonLat;
-  reprojectionObj *reprojectorToLonLat;
+  bool needsLonLat = false;
+  reprojectionObj *reprojectorToLonLat = nullptr;
 
-  mapObj
-      *mapToUseForWhichShapes; /* set if the map->extent and map->projection are
-                                  valid in msUVRASTERLayerWhichShapes() */
+  /* set if the map->extent and map->projection are
+     valid in msUVRASTERLayerWhichShapes() */
+  mapObj *mapToUseForWhichShapes = nullptr;
 
+  std::string timestring{};
+  std::string timefield{};
 } uvRasterLayerInfo;
+
+static uvRasterLayerInfo *getLayerInfo(layerObj *layer) {
+  return static_cast<uvRasterLayerInfo *>(layer->layerinfo);
+}
 
 void msUVRASTERLayerUseMapExtentAndProjectionForNextWhichShapes(layerObj *layer,
                                                                 mapObj *map) {
-  uvRasterLayerInfo *uvlinfo = (uvRasterLayerInfo *)layer->layerinfo;
+  uvRasterLayerInfo *uvlinfo = getLayerInfo(layer);
   uvlinfo->mapToUseForWhichShapes = map;
 }
 
 static int msUVRASTERLayerInitItemInfo(layerObj *layer) {
-  uvRasterLayerInfo *uvlinfo = (uvRasterLayerInfo *)layer->layerinfo;
+  uvRasterLayerInfo *uvlinfo = getLayerInfo(layer);
   int i;
   int *itemindexes;
   int failed = 0;
@@ -154,18 +160,13 @@ void msUVRASTERLayerFreeItemInfo(layerObj *layer) {
 }
 
 static void msUVRasterLayerInfoInitialize(layerObj *layer) {
-  uvRasterLayerInfo *uvlinfo = (uvRasterLayerInfo *)layer->layerinfo;
+  uvRasterLayerInfo *uvlinfo = getLayerInfo(layer);
 
   if (uvlinfo != NULL)
     return;
 
-  uvlinfo = (uvRasterLayerInfo *)msSmallCalloc(1, sizeof(uvRasterLayerInfo));
+  uvlinfo = new uvRasterLayerInfo;
   layer->layerinfo = uvlinfo;
-
-  uvlinfo->u = NULL;
-  uvlinfo->v = NULL;
-  uvlinfo->width = 0;
-  uvlinfo->height = 0;
 
   /* Set attribute type to Real, unless the user has explicitly set */
   /* something else. */
@@ -189,7 +190,7 @@ static void msUVRasterLayerInfoInitialize(layerObj *layer) {
 static void msUVRasterLayerInfoFree(layerObj *layer)
 
 {
-  uvRasterLayerInfo *uvlinfo = (uvRasterLayerInfo *)layer->layerinfo;
+  uvRasterLayerInfo *uvlinfo = getLayerInfo(layer);
 
   if (uvlinfo == NULL)
     return;
@@ -201,7 +202,7 @@ static void msUVRasterLayerInfoFree(layerObj *layer)
     msProjectDestroyReprojector(uvlinfo->reprojectorToLonLat);
   }
 
-  free(uvlinfo);
+  delete uvlinfo;
 
   layer->layerinfo = NULL;
 }
@@ -213,7 +214,7 @@ int msUVRASTERLayerOpen(layerObj *layer) {
   if (layer->layerinfo == NULL)
     return MS_FAILURE;
 
-  uvRasterLayerInfo *uvlinfo = (uvRasterLayerInfo *)layer->layerinfo;
+  uvRasterLayerInfo *uvlinfo = getLayerInfo(layer);
 
   uvlinfo->refcount = uvlinfo->refcount + 1;
 
@@ -227,7 +228,7 @@ int msUVRASTERLayerIsOpen(layerObj *layer) {
 }
 
 int msUVRASTERLayerClose(layerObj *layer) {
-  uvRasterLayerInfo *uvlinfo = (uvRasterLayerInfo *)layer->layerinfo;
+  uvRasterLayerInfo *uvlinfo = getLayerInfo(layer);
 
   if (uvlinfo != NULL) {
     uvlinfo->refcount--;
@@ -239,7 +240,7 @@ int msUVRASTERLayerClose(layerObj *layer) {
 }
 
 int msUVRASTERLayerGetItems(layerObj *layer) {
-  uvRasterLayerInfo *uvlinfo = (uvRasterLayerInfo *)layer->layerinfo;
+  uvRasterLayerInfo *uvlinfo = getLayerInfo(layer);
 
   if (uvlinfo == NULL)
     return MS_FAILURE;
@@ -273,7 +274,7 @@ static char **msUVRASTERGetValues(layerObj *layer, float u, float v,
   char tmp[100];
   float size_scale;
   int *itemindexes = (int *)layer->iteminfo;
-  uvRasterLayerInfo *uvlinfo = (uvRasterLayerInfo *)layer->layerinfo;
+  uvRasterLayerInfo *uvlinfo = getLayerInfo(layer);
   double lon = HUGE_VAL;
   double lat = HUGE_VAL;
 
@@ -452,7 +453,7 @@ rectObj msUVRASTERGetSearchRect(layerObj *layer, mapObj *map) {
 }
 
 int msUVRASTERLayerWhichShapes(layerObj *layer, rectObj rect, int isQuery) {
-  uvRasterLayerInfo *uvlinfo = (uvRasterLayerInfo *)layer->layerinfo;
+  uvRasterLayerInfo *uvlinfo = getLayerInfo(layer);
   imageObj *image_tmp;
   outputFormatObj *outputformat = NULL;
   mapObj *map_tmp;
@@ -753,7 +754,26 @@ int msUVRASTERLayerWhichShapes(layerObj *layer, rectObj rect, int isQuery) {
    */
   saved_layer_mask = layer->mask;
   layer->mask = NULL;
-  ret = msDrawRasterLayerLow(map_tmp, layer, image_tmp, NULL);
+
+  if (layer->tileindex) {
+    expressionObj old_filter;
+    if (!uvlinfo->timestring.empty()) {
+      msInitExpression(&old_filter);
+      msCopyExpression(&old_filter, &layer->filter); /* save existing filter */
+      msFreeExpression(&layer->filter);
+      msLayerMakeBackticsTimeFilter(layer, uvlinfo->timestring.c_str(),
+                                    uvlinfo->timefield.c_str());
+    }
+
+    ret = msDrawRasterLayerLow(map_tmp, layer, image_tmp, NULL);
+
+    if (!uvlinfo->timestring.empty()) {
+      msCopyExpression(&layer->filter, &old_filter); /* restore old filter */
+      msFreeExpression(&old_filter);
+    }
+  } else {
+    ret = msDrawRasterLayerLow(map_tmp, layer, image_tmp, NULL);
+  }
 
   /* restore layer attributes if we went through the above on-the-fly VRT */
   if (oldLayerData) {
@@ -797,6 +817,21 @@ int msUVRASTERLayerWhichShapes(layerObj *layer, rectObj rect, int isQuery) {
   uvlinfo->u = (float *)msSmallMalloc(sizeof(float *) * width * height);
   uvlinfo->v = (float *)msSmallMalloc(sizeof(float *) * width * height);
 
+  reprojectionObj *reprojectorLayerToMap = nullptr;
+  if (layer->project) {
+    reprojectorLayerToMap =
+        msProjectCreateReprojector(&layer->projection, &layer->map->projection);
+  }
+  // To reproject the wind direction, use a ~ 10 m long line. This is a value
+  // not too small and not too big to be able to reproject a small vector.
+  double typicalLength = 10.0;
+  if (msProjIsGeographicCRS(&(layer->projection))) {
+    const double dfDegToMeter =
+        msProjGetSemiMajorAxis(&(layer->projection)) * (MS_PI / 180);
+    const double dfMeterToDeg = 1.0 / dfDegToMeter;
+    typicalLength *= dfMeterToDeg;
+  }
+
   for (size_t off = 0; off < static_cast<size_t>(width) * height; ++off) {
     /* Ignore invalid pixels (at nodata), or (u,v)=(0,0) */
     if (MS_GET_BIT(image_tmp->img_mask, off)) {
@@ -805,6 +840,43 @@ int msUVRASTERLayerWhichShapes(layerObj *layer, rectObj rect, int isQuery) {
           image_tmp->img.raw_float[off + static_cast<size_t>(width) * height];
       if (!(uvlinfo->u[off] == 0 && uvlinfo->v[off] == 0)) {
         uvlinfo->query_results++;
+        if (reprojectorLayerToMap) {
+          // Compute coordinates of the pixel in layer projection
+          const int x = static_cast<int>(off % uvlinfo->width);
+          const int y = static_cast<int>(off / uvlinfo->width);
+          const double xLayer =
+              Pix2Georef(x, 0, uvlinfo->width - 1, uvlinfo->extent.minx,
+                         uvlinfo->extent.maxx, MS_FALSE);
+          const double yLayer =
+              Pix2Georef(y, 0, uvlinfo->height - 1, uvlinfo->extent.miny,
+                         uvlinfo->extent.maxy, MS_TRUE);
+
+          // Creates a vector starting at point.x,point.y, of length
+          // ~ 10 m, using the angle defined by (u,v)
+          pointObj point1, point2;
+          point1.x = xLayer;
+          point1.y = yLayer;
+          const double lengthInLayerProj =
+              hypotf(uvlinfo->u[off], uvlinfo->v[off]);
+          point2.x =
+              xLayer + typicalLength * (uvlinfo->u[off] / lengthInLayerProj);
+          point2.y =
+              yLayer + typicalLength * (uvlinfo->v[off] / lengthInLayerProj);
+
+          // Reprojects that vector to map projection
+          if (msProjectPointEx(reprojectorLayerToMap, &point1) == MS_SUCCESS &&
+              msProjectPointEx(reprojectorLayerToMap, &point2) == MS_SUCCESS) {
+            // Now computes the (u,v) component in map projection,
+            // preserving its original length
+            const double dx = point2.x - point1.x;
+            const double dy = point2.y - point1.y;
+            const double lengthInMapProj = hypot(dx, dy);
+            uvlinfo->u[off] =
+                (float)(lengthInLayerProj * (dx / lengthInMapProj));
+            uvlinfo->v[off] =
+                (float)(lengthInLayerProj * (dy / lengthInMapProj));
+          }
+        }
       } else {
         uvlinfo->u[off] = std::numeric_limits<float>::quiet_NaN();
         uvlinfo->v[off] = std::numeric_limits<float>::quiet_NaN();
@@ -814,6 +886,8 @@ int msUVRASTERLayerWhichShapes(layerObj *layer, rectObj rect, int isQuery) {
       uvlinfo->v[off] = std::numeric_limits<float>::quiet_NaN();
     }
   }
+
+  msProjectDestroyReprojector(reprojectorLayerToMap);
 
   msFreeImage(image_tmp); /* we do not need the imageObj anymore */
   msFreeMap(map_tmp);
@@ -825,7 +899,7 @@ int msUVRASTERLayerWhichShapes(layerObj *layer, rectObj rect, int isQuery) {
 
 int msUVRASTERLayerGetShape(layerObj *layer, shapeObj *shape,
                             resultObj *record) {
-  uvRasterLayerInfo *uvlinfo = (uvRasterLayerInfo *)layer->layerinfo;
+  uvRasterLayerInfo *uvlinfo = getLayerInfo(layer);
   lineObj line;
   pointObj point;
   const long shapeindex = record->shapeindex;
@@ -889,7 +963,7 @@ int msUVRASTERLayerGetShape(layerObj *layer, shapeObj *shape,
 }
 
 int msUVRASTERLayerNextShape(layerObj *layer, shapeObj *shape) {
-  uvRasterLayerInfo *uvlinfo = (uvRasterLayerInfo *)layer->layerinfo;
+  uvRasterLayerInfo *uvlinfo = getLayerInfo(layer);
 
   if (uvlinfo->next_shape < 0 ||
       uvlinfo->next_shape >= uvlinfo->query_results) {
@@ -1029,12 +1103,20 @@ int msUVRASTERLayerSetTimeFilter(layerObj *layer, const char *timestring,
   /* -------------------------------------------------------------------- */
   /*      If we are using a local shapefile as our tileindex (that is     */
   /*      to say, the tileindex name is not of another layer), then we    */
-  /*      just install a backtics style filter on the raster layer.       */
-  /*      This is propagated to the "working layer" created for the       */
-  /*      tileindex by code in mapraster.c.                               */
+  /*      will install a backtics style filter later.                     */
   /* -------------------------------------------------------------------- */
-  if (tilelayerindex == -1)
-    return msLayerMakeBackticsTimeFilter(layer, timestring, timefield);
+  if (tilelayerindex == -1) {
+    if (layer->layerinfo == NULL)
+      msUVRasterLayerInfoInitialize(layer);
+    if (layer->layerinfo == NULL)
+      return MS_FAILURE;
+    uvRasterLayerInfo *uvlinfo = getLayerInfo(layer);
+    if (timestring)
+      uvlinfo->timestring = timestring;
+    if (timefield)
+      uvlinfo->timefield = timefield;
+    return MS_SUCCESS;
+  }
 
   /* -------------------------------------------------------------------- */
   /*      Otherwise we invoke the tileindex layers SetTimeFilter          */
@@ -1047,7 +1129,7 @@ int msUVRASTERLayerSetTimeFilter(layerObj *layer, const char *timestring,
 }
 
 /************************************************************************/
-/*                msRASTERLayerInitializeVirtualTable()                 */
+/*                msUVRASTERLayerInitializeVirtualTable()               */
 /************************************************************************/
 
 int msUVRASTERLayerInitializeVirtualTable(layerObj *layer) {